# packages/credential-showcase-openapi/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Credential Showcase API
  version: 0.1.1
servers:
  - url: https://virtserver.swaggerhub.com/SphereonInt/showcase/0.1.0/v1
    description: SwaggerHub API Auto Mocking
  - url: https://api.example.com/v1
...
security:
  - OAuth2: []
  - OIDC: []
paths:
...
  /showcases:
    get:
      tags:
        - Showcases
      summary: List all showcases
      operationId: listShowcases
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShowcasesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Showcases
      summary: Create a new showcase
      operationId: createShowcase
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShowcaseRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowcaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /showcases/{showcaseId}:
    parameters:
      - name: showcaseId
        in: path
        required: true
        schema:
          type: string
        description: Unique identifier of the showcase
    get:
      tags:
        - Showcases
      summary: Get a specific showcase
      operationId: getShowcase
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowcaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Showcases
      summary: Update a showcase
      operationId: updateShowcase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShowcaseRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowcaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Showcases
      summary: Delete a showcase
      operationId: deleteShowcase
      responses:
        '204':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating the showcase was deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: <https://auth.example.com/oauth/authorize>
          tokenUrl: <https://auth.example.com/oauth/token>
          scopes:
            editor: Grants editor access
            admin: Grants admin access
            viewer: Grants viewer access
    OIDC:
      type: openIdConnect
      openIdConnectUrl: <https://auth.example.com/.well-known/openid-configuration>
      description: OpenID Connect security scheme for OAuth2 flows
  schemas:
    Asset:
      type: object
      required:
        - id
        - mediaType
        - content
      properties:
        id:
          type: string
          description: Unique identifier for the asset in the database
          example: 123e4567-e89b-12d3-a456-426614174469
        mediaType:
          type: string
          description: Type of asset (e.g., image, document, icon)
          example: image/jpeg
        content:
          type: string
          description: Binary data of the asset
          example: base64 encoded binary data
        fileName:
          type: string
          description: Name of the asset file
          example: asset.jpg
        description:
          type: string
          description: Description of the asset
          example: A beautiful image of a cat
...
    Step:
      type: object
      required:
        - id
        - title
        - description
        - order
        - type
        - actions
      properties:
        id:
          type: string
          description: Unique identifier for the step
          example: 123e4567-e89b-12d3-a456-434314174000
        title:
          type: string
          description: Title of the step
          example: Verify Identity
        description:
          type: string
          description: Detailed description of the step
          example: Verify the user's identity
        order:
          type: integer
          description: Order of the step in the workflow
          minimum: 0
          example: 1
        type:
          $ref: '#/components/schemas/StepType'
        subFlow:
          type: string
          description: Optional sub-workflow for this step
          example: 123e4567-e89b-12d3-a456-434314174000
        actions:
          type: array
          items:
            $ref: '#/components/schemas/StepAction'
        asset:
          $ref: '#/components/schemas/Asset'
...
    StepType:
      type: string
      enum:
        - HUMAN_TASK
        - SERVICE
        - WORKFLOW
      description: Type of step in the workflow
      example: HUMAN_TASK
    StepAction:
      type: object
      required:
        - id
        - actionType
        - title
        - text
      properties:
        id:
          type: string
          description: Unique identifier for the action
          example: 123e4567-ef2d-12d3-abcd-426614174456
        actionType:
          type: string
          description: Type of action
          example: EXAMPLE_ACTION
        title:
          type: string
          description: Title of the action
          example: Connect Wallet
        text:
          type: string
          description: Descriptive text for the action
          example: Connect your wallet to continue
      discriminator:
        propertyName: actionType
        mapping:
          ARIES_OOB: '#/components/schemas/AriesOOBAction'
...
    AriesOOBAction:
      allOf:
        - $ref: '#/components/schemas/StepAction'
        - type: object
          required:
            - proofRequest
          properties:
            proofRequest:
              $ref: '#/components/schemas/AriesProofRequest'
    AriesProofRequest:
      type: object
      properties:
        attributes:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AriesRequestCredentialAttributes'
          description: Map of attribute groups requested in the proof
          example:
            name:
              value: John Doe
        predicates:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AriesRequestCredentialPredicates'
          description: Map of predicate groups requested in the proof
          example:
            age:
              value: 30
    AriesRequestCredentialAttributes:
      type: object
      properties:
        attributes:
          type: array
          items:
            type: string
            description: Name of the attribute to be included in the proof
            example: name
          description: List of attribute names to be included in the proof
        restrictions:
          type: array
          items:
            type: string
            description: Name of the restriction to be included in the proof
            example: age
          description: List of restrictions on the requested attributes
    AriesRequestCredentialPredicates:
      type: object
      required:
        - name
        - type
        - value
        - restrictions
      properties:
        name:
          type: string
          description: Name of the attribute for predicate
          example: age
        type:
          type: string
          description: Type of predicate (e.g., >=, <=, etc.)
          example: '>='
        value:
          type: string
          description: Value to compare against
          example: '30'
        restrictions:
          type: array
          items:
            type: string
          description: List of restrictions on the predicate
    Persona:
      type: object
      required:
        - name
        - role
      properties:
        id:
          type: string
          description: Unique identifier for the persona
          example: 123e4567-e89b-12d3-a456-426614174456
        name:
          type: string
          description: Name of the persona
          example: John Doe
        role:
          type: string
          description: Role of the persona in the system
          example: Verifier
        description:
          type: string
          description: Detailed description of the persona
          example: John Doe is a verifier for the system
        headshotImage:
          $ref: '#/components/schemas/Asset'
        bodyImage:
          $ref: '#/components/schemas/Asset'
...
    Workflow:
      type: object
      required:
        - name
        - type
      properties:
        id:
          type: string
          description: Unique identifier for the workflow
          example: 789e4567-e89b-12d3-a456-434314174123
        name:
          type: string
          description: Name of the workflow
          example: Credential Issuance
        description:
          type: string
          description: Detailed description of the workflow
          example: This workflow issues credentials to users
        type:
          type: string
          enum: [ ISSUANCE, PRESENTATION ]
          description: Type of workflow
          example: ISSUANCE
        steps:
          type: array
          description: Ordered list of steps in the workflow
          items:
            $ref: '#/components/schemas/Step'
        personas:
          type: array
          description: References to personas used in this workflow
          items:
            $ref: '#/components/schemas/Persona'
...
    CredentialType:
      type: string
      enum:
        - ANONCRED
      description: Types of supported credentials
...
    CredentialDefinition:
      type: object
      required:
        - id
        - name
        - version
        - type
        - schemaId
        - icon
        - representations
      properties:
        id:
          type: string
          description: Unique identifier for the credential definition
          example: 123e4567-e89b-12d3-a456-426614174123
        name:
          type: string
          description: Name of the credential definition
          example: Credential Definition Name
        issuerId:
          type: string
          description: Unique identifier of the issuer who created this schema
          example: 123e4567-e89b-12d3-a456-426614174123
        schemaId:
          type: string
          description:  Identifier of the associated credential schema
          example:  123e4567-e89b-12d3-a456-426614174123
        version:
          type: string
          description: Version of the credential definition
          example: '1.0'
        type:
          $ref: '#/components/schemas/CredentialType'
        representations:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/CredentialRepresentation'
              - $ref: '#/components/schemas/OCARepresentation'
        revocation:
          oneOf:
            - $ref: '#/components/schemas/RevocationInfo'
            - $ref: '#/components/schemas/AnonCredRevocation'
        icon:
          $ref: '#/components/schemas/Asset'
...
    CredentialRepresentation:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: Unique identifier for the credential representation
          example: 123e4567-e89b-12d3-abcd-426614174456
    OCARepresentation:
      allOf:
        - $ref: '#/components/schemas/CredentialRepresentation'
        - type: object
          required:
            - credDefId
            - schemaId
          properties:
            credDefId:
              type: string
              description: Credential definition identifier in the OCA format
              example: 123e4567-e89b-12d3-a456-426614174123
            schemaId:
              type: string
              description: Schema identifier in the OCA format
              example: 123e4567-e89b-12d3-a456-426614174123
            ocaBundleUrl:
              type: string
              description: URL to the OCA bundle for this representation
              example: https://example.com/ocaBundle.json
      description: Overlay Capture Architecture representation of a credential
    RevocationInfo:
      type: object
      required:
        - title
      properties:
        id:
          type: string
          description: Id of the revocation information
          example: abcd4567-e89b-12d3-a456-426614174123
        title:
          type: string
          description: Title of the revocation information
          example: Revocation Information
        description:
          type: string
          description: Detailed description of the revocation process or requirements
          example: This credential is revocable
    AnonCredRevocation:
      allOf:
        - $ref: '#/components/schemas/RevocationInfo'
        - type: object
          required:
            - registryId
          properties:
            registryId:
              type: string
              description: Identifier of the revocation registry for AnonCreds
              example: 123e4567-e89b-12d3-a456-426614174123
      description: Anonymous Credentials specific revocation information
...
    Showcase:
      type: object
      required:
        - id
        - name
        - description
        - status
        - scenarios
        - personas
        - credentialDefinitions
        - hidden
      properties:
        id:
          type: string
          description: Unique identifier for the showcase
          example: 123e4567-e89b-12d3-a456-426614174456
        name:
          type: string
          description: Name of the showcase
          example: Credential Showcase BCGov
        description:
          type: string
          description: Description of the showcase
          example: Collection of credential usage scenarios
        status:
          $ref: '#/components/schemas/ShowcaseStatus'
        hidden:
          type: boolean
          description: Whether the showcase is hidden from public view
          default: false
        scenarios:
          type: array
          description: List of scenarios in this showcase
          items:
            $ref: '#/components/schemas/Workflow'
        credentialDefinitions:
          type: array
          description: List of credential definitions used in this showcase
          items:
            $ref: '#/components/schemas/CredentialDefinition'
        personas:
          type: object
          description: List of personas involved in this showcase
          items:
            $ref: '#/components/schemas/Persona'
    ShowcasesResponse:
      type: object
      properties:
        showcases:
          type: array
          description: "List of showcases"
          items:
            $ref: '#/components/schemas/Showcase'
    ShowcaseResponse:
      type: object
      properties:
        showcase:
          $ref: '#/components/schemas/Showcase'
    ShowcaseRequest:
      required:
        - name
        - description
        - status
        - scenarios
        - personas
        - credentialDefinitions
        - hidden
      properties:
        name:
          type: string
          description: Name of the showcase
          example: Credential Showcase BCGov
        description:
          type: string
          description: Description of the showcase
          example: Collection of credential usage scenarios
        status:
          $ref: '#/components/schemas/ShowcaseStatus'
        hidden:
          type: boolean
          description: Whether the showcase is hidden from public view
          default: false
        scenarios:
          type: array
          description: List of scenarios in this showcase
          items:
            type: string
            example: 123e4567-e89b-12d3-a456-426614174469
        credentialDefinitions:
          type: array
          description: List of credential definitions used in this showcase
          items:
            type: string
            example: 123e4567-e89b-12d3-a456-426614174469
        personas:
          type: array
          description: List of personas involved in this showcase
          items:
            type: string
            example: 123e4567-e89b-12d3-a456-426614174469
    ShowcaseStatus:
      type: string
      enum:
          - PENDING
          - ACTIVE
          - ARCHIVED
      description: Types of supported credentials

    # Scenario:
    #   type: object
    #   required:
    #     - name
    #     - description
    #   properties:
    #     id:
    #       type: string
    #       description: Unique identifier for the scenario
    #       example: 123e4567-e89b-12d3-a456-426614174456
    #     name:
    #       type: string
    #       description: Name of the scenario
    #       example: University Parking Journey
    #     description:
    #       type: string
    #       description: Description of the scenario
    #       example: Student obtaining and using parking credentials
    #     workflows:
    #       type: array
    #       description: References to workflows used in this scenario
    #       items:
    #         type: string
    #       example: ['123e4567-e89b-12d3-a456-426614174456', '123e4567-e89b-12d3-a456-426614174456']
    #     personas:
    #       type: array
    #       description: References to personas involved in this scenario
    #       items:
    #         type: string
    #       example: ['123e4567-e89b-12d3-a456-426614174456', '123e4567-e89b-12d3-a456-426614174456']
    # ScenarioRequest:
    #   type: object
    #   required:
    #     - name
    #     - description
    #   properties:
    #     name:
    #       type: string
    #       description: Name of the scenario
    #       example: University Parking Journey
    #     description:
    #       type: string
    #       description: Description of the scenario
    #       example: Student obtaining and using parking credentials
    #     workflows:
    #       type: array
    #       description: References to workflows used in this scenario
    #       items:
    #         type: string
    #       example: ['123e4567-e89b-12d3-a456-426614174456', '123e4567-e89b-12d3-a456-426614174456']
    #     personas:
    #       type: array
    #       description: References to personas involved in this scenario
    #       items:
    #         type: string
    #       example: ['123e4567-e89b-12d3-a456-426614174456', '123e4567-e89b-12d3-a456-426614174456']
    # ScenarioResponse:
    #   properties:
    #     assets:
    #       type: array
    #       items:
    #         $ref: '#/components/schemas/Scenario'
  responses:
    'BadRequest':
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Invalid request data'
    'Unauthorized':
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Invalid authentication credentials'
    'Forbidden':
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Insufficient permissions to access resource'
...
    'InternalServerError':
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'An unexpected error occurred'